test(core-p2p): refactor hapi-nes/plugin
