+++ __tests__/unit/core-p2p/hapi-nes/plugin.test.ts
